use int for qrcode + fix label def

+++ gui/qt/qrcodewidget.py
@@ -80,14 +80,14 @@ class QRCodeWidget(QWidget, util.PrintError):
         margin = 5
         framesize = min(r.width(), r.height())
         boxsize = int( (framesize - 2*margin)/k )
-        size = k*boxsize
-        left = (r.width() - size)/2
-        top = (r.height() - size)/2
+        size = int(k*boxsize)
+        left = int((r.width() - size)/2)
+        top = int((r.height() - size)/2)
 
         # Make a white margin around the QR in case of dark theme use
         qp.setBrush(self._white)
         qp.setPen(self._white)
-        qp.drawRect(left-margin, top-margin, size+(margin*2), size+(margin*2))
+        qp.drawRect(int(left-margin), int(top-margin), size+(int(margin*2)), size+(int(margin*2)))
         qp.setBrush(self._black)
         qp.setPen(self._black)
